requirements.txt
streamlit>=1.45.0
streamlit-autorefresh>=1.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
